Updated AccountSettings
Removed AccountSettingsPage.vue

AccountSettings.vue
- Page is working

student.js
- Updated GetStudentData function

index.ts
- Fixed AccountSettings
